commandHauls: extract rate limiting and user config to modules
