+++ src/Util/UserConfig.js
@@ -1,0 +1,5 @@
+import { Config } from '/twcheese/src/Models/Config.js';
+
+let userConfig = new Config('twcheese.userConfig');
+
+export { userConfig };
